refactor: api를 server 컴포넌트용이랑 client 컴포넌트용으로 분리
