+++ src/app/user-reports/page.tsx
@@ -1,4 +1,4 @@
-import { getUserReports } from "@/api/report";
+import { getUserReports } from "@/api/report.server";
 import Filter from "@/components/Filter";
 import Header from "@/components/Header";
 import Main from "@/components/Main";
